+++ src/main/java/com/uam/eCommerceUltimate/service/IServiceCategoria.java
@@ -10,9 +10,9 @@ public interface IServiceCategoria
 {
     public List<Categoria> getListCategoria();
 
-    public Categoria findById (int id);
+    public Categoria findById (Long id);
 
     public Categoria saveCategoria (Categoria categoria);
 
-    public void deleteCategoria(int id);
+    public void deleteCategoria(Long id);
 }
